# Optimise vertex buffer vertex additions