```diff
  | Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
  |------- |---------:|----------:|----------:|------:|------:|------:|----------:|
- |    Add | 249.0 ns |  1.528 ns |  1.429 ns |     - |     - |     - |         - |
+ |    Add | 42.88 ns | 0.2655 ns | 0.2483 ns |     - |     - |     - |         - |
```

I go from 120fps -> 200fps with chat opened to #lazer.